fix carousel
